Move chainLogos out of asset folder
