--- src/app/assets/chains/index.ts
@@ -1,15 +1,0 @@
-import babylon from "./babylon-genesis.png";
-import cosmos from "./cosmos.png";
-import ethereum from "./ethereum.png";
-import placeholder from "./placeholder.svg";
-import sui from "./sui.png";
-
-export const chainLogos = {
-  babylon,
-  cosmos,
-  ethereum,
-  sui,
-  placeholder,
-} as const;
-
-export { babylon, cosmos, ethereum, sui, placeholder };
