+++ src/app/components/Multistaking/MultistakingForm/FormValuesConsumer.tsx
@@ -2,8 +2,8 @@ import { useFormContext, useWatch } from "@babylonlabs-io/core-ui";
 import Image from "next/image";
 import { useEffect } from "react";
 
-import { chainLogos } from "@/app/assets/chains";
 import { MultistakingPreviewModal } from "@/app/components/Modals/MultistakingModal/MultistakingStartModal";
+import { chainLogos } from "@/app/components/Multistaking/chainLogos";
 import { useStakingService } from "@/app/hooks/services/useStakingService";
 import type { FormFields } from "@/app/state/StakingState";
 import { useStakingState } from "@/app/state/StakingState";
